moved DisjointSet to _lib module
