+++ scipy/_lib/_disjoint_set.py
