+++ scipy/cluster/__init__.py
@@ -20,9 +20,10 @@ calculating statistics on clusters, cutting linkages
 to generate flat clusters, and visualizing clusters with dendrograms.
 
 """
-__all__ = ['vq', 'hierarchy']
+__all__ = ['vq', 'hierarchy', 'DisjointSet']
 
 from . import vq, hierarchy
+from scipy._lib._disjoint_set import DisjointSet
 
 from scipy._lib._testutils import PytestTester
 test = PytestTester(__name__)
