+++ scipy/sparse/csgraph/__init__.py
@@ -191,7 +191,6 @@ from ._reordering import reverse_cuthill_mckee, structural_rank
 from ._tools import construct_dist_matrix, reconstruct_path,\
     csgraph_from_dense, csgraph_to_dense, csgraph_masked_from_dense,\
     csgraph_from_masked, csgraph_to_masked
-from ._disjoint_set import DisjointSet
 
 from scipy._lib._testutils import PytestTester
 test = PytestTester(__name__)
